Add distribution validator
